gltf: Move animation processing out of gltfpack.cpp
