+++ CMakeLists.txt
@@ -31,6 +31,7 @@ set(SOURCES
 )
 
 set(GLTF_SOURCES
+    gltf/animation.cpp
     gltf/basistoktx.cpp
     gltf/parseobj.cpp
     gltf/parsegltf.cpp
