+++ gltf/gltfpack.h
@@ -183,6 +183,8 @@ struct BufferView
 	size_t bytes;
 };
 
+void processAnimation(Animation& animation, const Settings& settings);
+
 cgltf_data* parseObj(const char* path, std::vector<Mesh>& meshes, const char** error);
 cgltf_data* parseGltf(const char* path, std::vector<Mesh>& meshes, std::vector<Animation>& animations, const char** error);
